+++ source/extensions/isaacsim.asset.importer.urdf.ui/python/tests/test_ui.py
@@ -135,7 +135,9 @@ class TestImporterUI(MenuUITestCase):
         await ui_test.human_delay()
         await omni.kit.app.get_app().next_update_async()
 
-        grid_view = await self.find_widget_with_retry("Select File//Frame/**/VGrid[*].identifier=='None_grid_view'")
+        grid_view = await self.find_widget_with_retry(
+            "Select File//Frame/**/VGrid[*].identifier=='filebrowser_grid_view'"
+        )
         file = await self.find_widget_with_retry("**/Label[*].text=='ur10.urdf'", parent=grid_view)
         await file.click()
         await ui_test.human_delay()
@@ -195,7 +197,9 @@ class TestImporterUI(MenuUITestCase):
         await ui_test.human_delay()
         await omni.kit.app.get_app().next_update_async()
 
-        grid_view = await self.find_widget_with_retry("Select File//Frame/**/VGrid[*].identifier=='None_grid_view'")
+        grid_view = await self.find_widget_with_retry(
+            "Select File//Frame/**/VGrid[*].identifier=='filebrowser_grid_view'"
+        )
         file = await self.find_widget_with_retry("**/Label[*].text=='ur10.urdf'", parent=grid_view)
         await file.click()
         await ui_test.human_delay()