+++ source/extensions/omni.kit.loop-isaac/bindings/BindingsRunLoopRunnerPython.cpp
@@ -94,17 +94,6 @@ PYBIND11_MODULE(_loop, m)
                     :obj:`double`: The time step size in seconds.
 
                 )pbdoc",
-             py::arg("name") = "")
-        .def("set_next_simulation_time", wrapInterfaceFunction(&IRunLoopRunnerImpl::setNextSimulationTime),
-             R"pbdoc(
-                Sets the next simulation time to be picked up by the run loop.
-
-                Args:
-                    time (:obj:`double`): The simulation time value in seconds. Use negative value to reset to frame-based time.
-
-                    name (:obj:`str`): The name of the run loop. If name is an empty string, all active run loops are set.
-
-                )pbdoc",
-             py::arg("time") = 0.0, py::arg("name") = "");
+             py::arg("name") = "");
 }
 }